HTCONDOR-2021: Rename module to ansi without two I's
